# docs: benchmark