--- a/.eslintignore
+++ b/.eslintignore
@@ -1,1 +1,2 @@
 src/types/global.d.ts
+benchmark/