--- a/benchmark/.swcrc
+++ b/benchmark/.swcrc
@@ -1,0 +1,13 @@
+{
+  "env": {
+    "targets": "node >= 16"
+  },
+  "jsc": {
+    "target": "es2022",
+    "parser": {
+      "syntax": "typescript",
+      "tsx": false,
+      "dynamicImport": true
+    }
+  }
+}