--- a/benchmark/package.json
+++ b/benchmark/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "benchmark",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "build": "swc src/samples --out-dir dist",
+    "bench": "npm run build && node dist/samples/compare-libraries.js"
+  },
+  "type": "module",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/H4ad/serverless-adapter.git"
+  },
+  "license": "MIT",
+  "author": {
+    "name": "Vinícius Lourenço",
+    "email": "H4ad@users.noreply.github.com",
+    "url": "https://github.com/H4ad"
+  },
+  "dependencies": {
+    "@h4ad/serverless-adapter": "file:../h4ad-serverless-adapter-0.0.0-development.tgz",
+    "@swc/cli": "^0.1.59",
+    "@swc/core": "^1.3.24",
+    "@vendia/serverless-express": "^4.10.1",
+    "cronometro": "^1.1.5",
+    "express": "^4.18.2",
+    "serverless-http": "^3.1.0",
+    "stream-mock": "^2.0.5"
+  },
+  "devDependencies": {
+    "aws-lambda": "^1.0.7",
+    "chokidar": "^3.5.3",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.4"
+  }
+}