--- a/tsconfig.eslint.json
+++ b/tsconfig.eslint.json
@@ -1,5 +1,5 @@
 {
   "extends": "./tsconfig.json",
   "include": ["src/**/*.ts", "test/**/*.ts", "scripts/**/*.ts", "src/index.doc.ts"],
-  "exclude": []
+  "exclude": ["benchmark/**/*.ts"]
 }