<!--
  😀 Wonderful!  Thank you for opening a pull request.

  Please fill in the information below to expedite the review
  and (hopefully) merge of your change.
-->

### Description of change

Added benchmark tests to check if the library is performing good.

Also, I improve the performance by 100% by just deferring the logging that uses `util.inspect`.
This function takes a lot of time to execute and is very heavy, if the logging library is not configured to log
that information, don't worth paying the runtime execution cost to log that information.

In the future, I will add another way of logging but for now, the solution of deferring the execution by logging a function will work.

The old stack trace:

![serverless-adapter-old](https://user-images.githubusercontent.com/12551007/210153330-24874c32-ef6f-4d44-990c-74dd1b39d165.png)

> The `onReceiveRequest` functions took 3ms of 9ms execution time.

The new stack trace:

![serverless-adapter-new](https://user-images.githubusercontent.com/12551007/210153372-89483690-dbcb-41b8-a19d-145259f2c6fa.png)

Before the fix:

```md
╔════════════════════════════╤═════════╤════════════════╤═══════════╗
║ Slower tests               │ Samples │         Result │ Tolerance ║
╟────────────────────────────┼─────────┼────────────────┼───────────╢
║ @h4ad/serverless-adapter   │    5500 │ 4424.96 op/sec │  ± 0.98 % ║
║ @vendia/serverless-express │    6500 │ 4454.06 op/sec │  ± 0.97 % ║
╟────────────────────────────┼─────────┼────────────────┼───────────╢
║ Fastest test               │ Samples │         Result │ Tolerance ║
╟────────────────────────────┼─────────┼────────────────┼───────────╢
║ serverless-http            │   10000 │ 8046.93 op/sec │  ± 1.48 % ║
╚════════════════════════════╧═════════╧════════════════╧═══════════╝
```

After the fix:

```md
╔════════════════════════════╤═════════╤════════════════╤═══════════╗
║ Slower tests               │ Samples │         Result │ Tolerance ║
╟────────────────────────────┼─────────┼────────────────┼───────────╢
║ @vendia/serverless-express │    5500 │ 4518.75 op/sec │  ± 0.99 % ║
║ @h4ad/serverless-adapter   │   10000 │ 7590.44 op/sec │  ± 1.29 % ║
╟────────────────────────────┼─────────┼────────────────┼───────────╢
║ Fastest test               │ Samples │         Result │ Tolerance ║
╟────────────────────────────┼─────────┼────────────────┼───────────╢
║ serverless-http            │   10000 │ 8298.48 op/sec │  ± 1.53 % ║
╚════════════════════════════╧═════════╧════════════════╧═══════════╝
```

<!--
  Please be clear and concise what the change is intended to do,
  why this change is needed, and how you've verified that it
  corrects what you intended.

  In some cases it may be helpful to include the current behavior
  and the new behavior.

  If the change is related to an open issue, you can link it here.
  If you include `Fixes #0000` (replacing `0000` with the issue number)
  when this is merged it will automatically mark the issue as fixed and
  close it.
-->

### Pull-Request Checklist

<!--
  Please make sure to review and check all of the following.

  If an item is not applicable, you can add "N/A" to the end.
-->

- [x] Code is up-to-date with the `main` branch
- [x] `npm run lint` passes with this change
- [x] `npm run test` passes with this change
- [x] This pull request links relevant issues as `Fixes #0000`
- [x] There are new or updated unit tests validating the change
- [x] Added documentation inside `www/docs/main` folder.
- [x] Included new files inside `index.doc.ts`.
- [x] The new commits follow conventions outlined in the [conventional commit spec](https://www.conventionalcommits.org/en/v1.0.0/)

<!--
  🎉 Thank you for contributing!
-->